Gameplay mechanics / Cinemachine / 2D rendering  (#9)
* Refactored Mech and Player Controller code

* Set Mech Controller script to be disabled by default

* Added name spacing

* Updated mech and player prefabs and their scripts

Scale of Mech is now twice the size of the player. Fixed camera jitter by setting both rigidbodies to "continuous detection" and turned "Interpolation" on. Both scripts are have an offset and ground check radius set to default values. During run time, the mech script has logic to override the ground check radius to accommodate the mechs scale.

* Created Camera prefab

* Set "Main Camera" tag on main camera

* Update Cameras.prefab

Can't preset a follow or look at target. It has to be none, or it will cause weird issues.

* Cleaned up input check logic

* Created Launch player mechanic

* Created asset files for 2D rendering and post-processing