Added different types for dialogue
Different types are:
Bubble-Will create a text bubble over the speakers head (Still needs to be created)
TextBox-Creates a textbox on the bottom of the screen, need to implement that this fires on an interaction key being pressed (E?), and stops all movement controls on the player.
InteractableTextBox-Same as above, but will have a list of buttons (Yes/No, Ok/Cancel) to provide for multiple choices.