Modified door and lever scripts
Lever script now has a selection between whether or not it interacts with doors/mirrors.
If mirrors, lever still interacts as normal.
If doors, the door now has an ID parameter, stored as a string. The lever will broadcast out an ID to one that matches, and if one is found, it will open/close that door.